% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_speed.R
\name{flow_speed}
\alias{flow_speed}
\title{Flow-speed relationship}
\usage{
flow_speed(Sac1_flow, reach)
}
\arguments{
\item{Sac1_flow}{Flow in Sac1 reach}

\item{reach}{One of following reaches: GeoDCC, Sac1, Sac2, Verona_to_Sac}
}
\description{
Calculates migration speed (km/day) as a function of Sac1 flow (cfs)
}
\examples{
flow_speed(1e3, "Sac1")
flow_speed(1e4, "Sac1")
flow_speed(1e5, "Sac1")
flow_speed(1e4, "GeoDCC")
}
