% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_survival.R
\name{flow_survival}
\alias{flow_survival}
\title{Flow-survival relationship}
\usage{
flow_survival(reach = c("Sac1", "Sac2", "Sac3", "SS"), flow)
}
\arguments{
\item{reach}{Sac1, Sac2, Sac3, or SS}

\item{flow}{Standardized flow in Sac3 on day entering reach}
}
\description{
Calculates survival through a reach as a function of standardized flow in Sac3 reach
}
\examples{
flow_survival("Sac1", -5)
flow_survival("Sac1", 0)
flow_survival("Sac1", 5)
flow_survival("SS", 0)

}
