% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interior_delta_survival.R
\name{interior_delta_survival}
\alias{interior_delta_survival}
\title{Interior Delta survival}
\usage{
interior_delta_survival(n, scenario, model_day, flow_list)
}
\arguments{
\item{n}{GeoDCC, Sac1, Sac2, Sac3, Sac4, SS, Verona_to_Sac, Yolo, or Interior Delta}

\item{scenario}{Scenario describing water project operations}

\item{model_day}{Integer day in 82-yr model period}

\item{flow_list}{List that contains flow values for all reaches, scenarios, and days in 82-yr model period}
}
\description{
Draw random value for survival through the Interior Delta on a given day.
}
\examples{

}
