% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reach_length}
\alias{reach_length}
\title{Reach length}
\format{A named vector of lengths for the following reaches:
\describe{
  \item{GeoDCC}{Combined reach of Georgiana Slough, Delta Cross Channel, and South and North Forks
  of the Mokelumne River ending at confluence with the San Joaquin River.}
  \item{Sac1}{Sacramento River from Freeport to junction with Sutter Slough.}
  \item{Sac2}{Sacramento River from Sutter Slough junction to junction with Delta Cross Channel.}
  \item{Sac3}{Sacramento River from Delta Cross Channel junction to Rio Vista.}
  \item{Sac4}{Sacramento River from Rio Vista to Chipps Island.}
  \item{SS}{Combined reach of Sutter Slough and Steamboat Slough ending at Rio Vista.}
  \item{Verona_to_Sac}{Sacramento River from Verona to Freeport.}
}}
\usage{
reach_length
}
\description{
Length (km) of reaches in the Delta Passage Model.
}
\keyword{datasets}
