% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{routing}
\alias{routing}
\title{Routing}
\usage{
routing(reach, scenario, model_day, flow_list)
}
\arguments{
\item{reach}{GeoDCC, Sac1, Sac2, Sac3, Sac4, SS, Verona_to_Sac, Yolo, or Interior Delta}

\item{scenario}{Scenario describing water project operations}

\item{model_day}{Integer day in 82-yr model period}

\item{flow_list}{List that contains flow values for all reaches, scenarios, and days in 82-yr model period}
}
\description{
Determine next reach(es) for a cohort based on current reach.
}
\examples{
routing("Sac1", "EXG", 2, list("SS" = list("EXG" = c(1000, 5000, 10000)), "Sac2" = list("EXG" = c(2000, 10000, 20000))))
routing("SS", "EXG", 2, list("SS" = list("EXG" = c(1000, 5000, 10000)), "Sac2" = list("EXG" = c(2000, 10000, 20000))))

}
