% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtruncnorm_surv.R
\name{rtruncnorm_surv}
\alias{rtruncnorm_surv}
\title{Truncated normal survival}
\usage{
rtruncnorm_surv(n, mean, sd)
}
\arguments{
\item{n}{Number of random values to draw from truncted normal distribution}

\item{mean}{Mean of truncated normal distribution}

\item{sd}{Standard deviation of truncated normal distribution}
}
\description{
Wrapper function for rtruncnorm where parameters a and b equal 0 and 1, respectively.
}
