% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sac_timing}
\alias{sac_timing}
\title{Sacramento River entry timing}
\format{A list with 5 vectors of 29950 elements:
\describe{
  \item{Date}{character string representing the date}
  \item{Winter}{winter-run daily entry proportion}
  \item{Spring}{spring-run daily entry proportion}
  \item{Fall}{fall-run daily entry proportion}
  \item{LateFall}{late-fall-run daily entry proportion}
}}
\usage{
sac_timing
}
\description{
A dataset containing the proportion of the population that enters the model
in the Sacramento River at Freeport on each day of the 82-yr model period.
}
\keyword{datasets}
