% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cohort.R
\name{sim_cohort}
\alias{sim_cohort}
\title{Simulate cohort}
\usage{
sim_cohort(run, day, reach, abundance)
}
\arguments{
\item{run}{Chinook Salmon run}

\item{day}{Model day that cohort is entering the model}

\item{reach}{Reach where cohort enters the model}

\item{abundance}{Number of fish in cohort when entering the model}
}
\description{
Simulate cohort as it passes through the Delta
}
\examples{


}
