% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{Survival}
\usage{
survival(reach, scenario, model_day, flow_list)
}
\arguments{
\item{reach}{GeoDCC, Sac1, Sac2, Sac3, Sac4, SS, Verona_to_Sac, Yolo, or Interior Delta}

\item{scenario}{Scenario describing water project operations}

\item{model_day}{Integer day in 82-yr model period}

\item{flow_list}{List that contains flow values for all reaches, scenarios, and days in 82-yr model period}
}
\description{
Draw random value for survival through a reach on a given day.
}
\examples{
survival("Sac3", "EXG", 2, list("Sac3" = list("Standardized" = list("EXG" = c(-5, 0, 5)))))
survival("Sac2", "EXG", 2, list("Sac3" = list("Standardized" = list("EXG" = c(-5, 0, 5)))))

}
