% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cohort.R
\name{update_cohort}
\alias{update_cohort}
\title{Update cohort}
\usage{
update_cohort(cohort_list, reach, day, abundance, speed, survival)
}
\arguments{
\item{cohort_list}{List of cohort parameters}

\item{reach}{Reach that cohort is entering}

\item{day}{Model day that cohort is entering reach}

\item{abundance}{Number of fish in cohort}

\item{speed}{Daily migration rate of cohort}

\item{survival}{Survival rate of cohort}
}
\description{
Update list of cohort parameters
}
\examples{
cohort <- initialize_cohort("Winter", "Sac1", 1, 1e6, 10, 0.8)
update_cohort(cohort, "Sac2", 10, 1e5, 15, 0.9)
}
