% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_dpm_structure.R
\name{view_dpm_structure}
\alias{view_dpm_structure}
\title{Create pkgnet package report}
\usage{
view_dpm_structure()
}
\description{
Create pkgnet package report that launches in default browser. The 'Function Network' tab is useful for visualizing
how DPM is structured by showing a network graph of the functions.
}
