% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_speed.R
\name{flow_speed}
\alias{flow_speed}
\title{Flow-speed relationship}
\usage{
flow_speed(reach, flow)
}
\arguments{
\item{reach}{GeoDCC, Sac1, Sac2, or Verona_to_Sac}

\item{flow}{Flow on day entering reach}
}
\description{
Calculates migration speed (km/day) as a function of flow (cfs)
}
\examples{
flow_speed("Sac1", 1e3)
flow_speed("Sac1", 1e4)
flow_speed("Sac1", 1e5)
flow_speed("GeoDCC", 1e4)
}
