% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_survival_sd.R
\docType{data}
\name{flow_survival_sd}
\alias{flow_survival_sd}
\title{Survival standard deviation}
\format{A list of 4 data frames with StandardizedFlow and SurvivalSD columns:
\describe{
  \item{Sac1}{Sacramento River from Freeport to junction with Sutter Slough.}
  \item{Sac2}{Sacramento River from Sutter Slough junction to junction with Delta Cross Channel.}
  \item{Sac3}{Sacramento River from Delta Cross Channel junction to Rio Vista, CA.}
  \item{SS}{Combined reach of Sutter Slough and Steamboat Slough ending at Rio Vista, CA.}
}}
\usage{
flow_survival_sd
}
\description{
Look-up tables for the standard deviation in survival for a given standardized flow.
Tables are used to apply stochasticity to survival in reaches that use a flow-survival relationship.
}
\keyword{datasets}
