% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_cohort.R
\name{initialize_cohort}
\alias{initialize_cohort}
\title{Initialize cohort}
\usage{
initialize_cohort(run, reach, day, abundance, travel_time, survival)
}
\arguments{
\item{run}{Chinook Salmon run}

\item{reach}{Reach that cohort is entering}

\item{day}{Model day that cohort is entering reach}

\item{abundance}{Number of fish in cohort}

\item{travel_time}{Travel time for cohort through reach}

\item{survival}{Survival rate of cohort through reach}
}
\description{
Initialize list of cohort parameters
}
\examples{
initialize_cohort("Winter", "Sac1", 1, 1e6, 10, 0.8)

}
