% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cohort.R
\name{sim_cohort}
\alias{sim_cohort}
\title{Simulate cohort}
\usage{
sim_cohort(abundance, scenario, model_day, flow_list)
}
\arguments{
\item{abundance}{Number of fish in cohort when entering the model}

\item{scenario}{Scenario describing water project operations}

\item{model_day}{Integer day in 82-yr model period}

\item{flow_list}{List that contains flow values for all reaches, scenarios, and days in 82-yr model period}
}
\description{
Simulate cohort as it passes through the Delta
}
